Implement the new "view.delta" command replacing "view.changes"
Renamed query constants and methods.
Renamed and updated classes that wrap options and results.
Fixed signatures on related extension methods.

Added special handling of the 410 error coming from a delta response,
swallowing the ODServerException and returning an actionable response
object that can be used to query for the new result.
The implementation of handling this error is a bit awkward, and doesn't
really fit with the patterns of the rest of the library. However, the
protocol doesn't behave this way for any other query either.
It was done this way to minimize impact and make it easiest to handle from
calling code.